+++ Clicked/Core/Database.lua
@@ -348,6 +348,7 @@ function Addon:GetNewBindingTemplate()
 			instanceType = GetTriStateLoadOptionTemplate("NONE"),
 			zoneName = GetLoadOptionTemplate(""),
 			equipped = GetLoadOptionTemplate(""),
+			bonusbar = GetLoadOptionTemplate(""),
 			channeling = GetNegatableStringLoadOptionTemplate(),
 			flying = GetNegatableLoadOptionTemplate(),
 			flyable = GetNegatableLoadOptionTemplate(),