+++ Clicked/Definitions.lua
@@ -105,6 +105,7 @@
 --- @field public pvpTalent Binding.TriStateLoadOption
 --- @field public warMode Binding.LoadOption
 --- @field public channeling Binding.NegatableStringLoadOption
+--- @field public bonusbar Binding.NegatableStringLoadOption
 
 --- @class Binding.LoadOption
 --- @field public selected boolean
@@ -151,6 +152,7 @@
 --- @field public swimming string
 --- @field public channeling Action.NegatableValueString
 --- @field public forms Action.NegatableValueString
+--- @field public bonusbar Action.NegatableValueString
 --- @field public unit string
 --- @field public hostility string
 --- @field public vitals string