+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Use new import and export window for profile import, export, and direct player share
 * Improve warning shown when the self-cast modifier is enabled
+* Add `bonusbar` and `nobonusbar` macro conditions
 
 ### Fixed